refactor(component): Integrate skeleton components
