+++ apps/web/components/portfolio/card.tsx
@@ -5,6 +5,7 @@ import {
   CardFooter,
   CardHeader,
   CardTitle,
+  Skeleton,
 } from '@juun/ui';
 import Image from 'next/image';
 import Link from 'next/link';
@@ -81,5 +82,37 @@ export function PortfolioCard({ post }: PortfolioCardProps) {
         </CardFooter>
       </Card>
     </Link>
+  );
+}
+
+export function PortfolioCardSkeleton() {
+  return (
+    <Card className="h-full overflow-hidden">
+      {/* Image skeleton */}
+      <Skeleton className="aspect-video w-full" />
+
+      {/* Title and description skeleton */}
+      <CardHeader className="space-y-2">
+        <Skeleton className="h-6 w-3/4" />
+        <Skeleton className="h-4 w-full" />
+        <Skeleton className="h-4 w-full" />
+        <Skeleton className="h-4 w-2/3" />
+      </CardHeader>
+
+      {/* Tags skeleton */}
+      <CardContent>
+        <div className="flex flex-wrap gap-2">
+          <Skeleton className="h-5 w-16 rounded-full" />
+          <Skeleton className="h-5 w-20 rounded-full" />
+          <Skeleton className="h-5 w-14 rounded-full" />
+        </div>
+      </CardContent>
+
+      {/* Footer skeleton */}
+      <CardFooter className="flex items-center justify-between">
+        <Skeleton className="h-4 w-24" />
+        <Skeleton className="h-3 w-16" />
+      </CardFooter>
+    </Card>
   );
 }
